feat: expose util functions
