+++ packages/ogre/src/git2json.ts
@@ -1,7 +1,11 @@
 import { Commit } from "./commit";
 import { History, Reference } from "./interfaces";
-import { createHeadRefValue, isTagRef, REFS_HEAD_KEY } from "./repository";
-import { cleanAuthor } from "./utils";
+import {
+  cleanAuthor,
+  createHeadRefValue,
+  isTagRef,
+  REFS_HEAD_KEY,
+} from "./utils";
 
 const findRefs = (commit: Commit, refs: Map<string, Reference>) => {
   const list = [];
